requirements.txt
flask
flask-login
cryptography
python-dotenv
